Clean up InvocationMatcher functionality
