+++ src/main/java/org/openrewrite/analysis/InvocationMatcher.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 the original author or authors.
+ * Copyright 2023 the original author or authors.
  * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,6 +18,7 @@ package org.openrewrite.analysis;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import org.openrewrite.Cursor;
+import org.openrewrite.Incubating;
 import org.openrewrite.internal.lang.Nullable;
 import org.openrewrite.java.MethodMatcher;
 import org.openrewrite.java.tree.*;
@@ -26,17 +27,39 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 /**
- * Matcher for finding {@link J.NewClass} and {@link J.MethodInvocation} {@link Expression}s.
+ * The most basic version of a {@link MethodMatcher} that allows implementers to craft custom matching logic.
  */
+@Incubating(since = "8.1.3")
 @FunctionalInterface
 public interface InvocationMatcher {
-    boolean matches(@Nullable Expression expression);
 
-    default AdvancedInvocationMatcher advanced() {
-        return new AdvancedInvocationMatcher(this);
+    /**
+     * Whether the method invocation or constructor matches the criteria of this matcher.
+     *
+     * @param type The type of the method invocation or constructor.
+     * @return True if the invocation or constructor matches the criteria of this matcher.
+     */
+    boolean matches(@Nullable JavaType.Method type);
+
+    default boolean matches(@Nullable MethodCall methodCall) {
+        if (methodCall == null) {
+            return false;
+        }
+        return matches(methodCall.getMethodType());
+    }
+
+    /**
+     * Whether the method invocation or constructor matches the criteria of this matcher.
+     *
+     * @param maybeMethod Any {@link Expression} that might be a method invocation or constructor.
+     * @return True if the invocation or constructor matches the criteria of this matcher.
+     */
+    default boolean matches(@Nullable Expression maybeMethod) {
+        return maybeMethod instanceof MethodCall && matches(((MethodCall) maybeMethod).getMethodType());
     }
 
     static InvocationMatcher from(Collection<? extends InvocationMatcher> matchers) {
@@ -49,38 +72,45 @@ public interface InvocationMatcher {
         return expression -> matchers.stream().anyMatch(matcher -> matcher.matches(expression));
     }
 
-    static InvocationMatcher fromMethodMatchers(Collection<? extends MethodMatcher> matchers) {
-        return from(matchers.stream().map(InvocationMatcher::fromMethodMatcher).collect(Collectors.toSet()));
+    static InvocationMatcher fromMethodMatcher(MethodMatcher methodMatcher) {
+        return methodMatcher::matches;
     }
 
     static InvocationMatcher fromMethodMatchers(MethodMatcher... matchers) {
         return fromMethodMatchers(Arrays.asList(matchers));
     }
 
-    static InvocationMatcher fromMethodMatcher(MethodMatcher matcher) {
-        return matcher::matches;
+    static InvocationMatcher fromMethodMatchers(Collection<? extends MethodMatcher> matchers) {
+        return from(matchers.stream().map(InvocationMatcher::fromMethodMatcher).collect(Collectors.toSet()));
+    }
+
+    default Advanced advanced() {
+        return new Advanced(this);
     }
 
     @AllArgsConstructor(access = AccessLevel.PRIVATE)
-    class AdvancedInvocationMatcher {
+    final class Advanced {
         private InvocationMatcher matcher;
 
         public boolean isSelect(Cursor cursor) {
-            Expression expression = ensureCursorIsExpression(cursor);
-            assert expression == cursor.getValue() : "expression != cursor.getValue()";
-            J.MethodInvocation maybeMethodInvocation =
-                    cursor.getParentOrThrow().firstEnclosing(J.MethodInvocation.class);
-            return maybeMethodInvocation != null &&
-                   maybeMethodInvocation.getSelect() == expression &&
-                   matcher.matches(maybeMethodInvocation); // Do the matcher.matches(...) last as this can be expensive
+            return asExpression(cursor, expression -> {
+                assert expression == cursor.getValue() : "expression != cursor.getValue()";
+                J.MethodInvocation maybeMethodInvocation =
+                        cursor.getParentOrThrow().firstEnclosing(J.MethodInvocation.class);
+                return maybeMethodInvocation != null &&
+                        maybeMethodInvocation.getSelect() == expression &&
+                        matcher.matches(maybeMethodInvocation); // Do the matcher.matches(...) last as this can be expensive
+            });
         }
 
         public boolean isAnyArgument(Cursor cursor) {
-            Expression expression = ensureCursorIsExpression(cursor);
-            return nearestMethodCall(cursor).map(call -> {
-                return call.getArguments().contains(expression)
-                       && matcher.matches(call); // Do the matcher.matches(...) last as this can be expensive
-            }).orElse(false);
+            return asExpression(
+                    cursor,
+                    expression -> nearestMethodCall(cursor).map(call -> {
+                        return call.getArguments().contains(expression)
+                                && matcher.matches(call); // Do the matcher.matches(...) last as this can be expensive
+                    }).orElse(false)
+            );
         }
 
         public boolean isFirstParameter(Cursor cursor) {
@@ -88,11 +118,10 @@ public interface InvocationMatcher {
         }
 
         public boolean isParameter(Cursor cursor, int parameterIndex) {
-            Expression expression = ensureCursorIsExpression(cursor);
             if (parameterIndex < 0) {
                 throw new IllegalArgumentException("parameterIndex < 0");
             }
-            return nearestMethodCall(cursor).map(call -> {
+            return asExpression(cursor, expression -> nearestMethodCall(cursor).map(call -> {
                 List<Expression> arguments = call.getArguments();
                 if (parameterIndex >= arguments.size()) {
                     return false;
@@ -109,12 +138,12 @@ public interface InvocationMatcher {
                     if (finalParameterIndex == parameterIndex) {
                         List<Expression> varargs = arguments.subList(finalParameterIndex, arguments.size());
                         return varargs.contains(expression) &&
-                               matcher.matches(call); // Do the matcher.matches(...) last as this can be expensive
+                                matcher.matches(call); // Do the matcher.matches(...) last as this can be expensive
                     }
                 }
                 return arguments.get(parameterIndex) == expression &&
-                       matcher.matches(call); // Do the matcher.matches(...) last as this can be expensive
-            }).orElse(false);
+                        matcher.matches(call); // Do the matcher.matches(...) last as this can be expensive
+            }).orElse(false));
         }
 
         private static boolean doesMethodHaveVarargs(MethodCall expression) {
@@ -133,12 +162,8 @@ public interface InvocationMatcher {
             return Optional.empty();
         }
 
-        private static Expression ensureCursorIsExpression(Cursor cursor) {
-            if (cursor.getValue() instanceof Expression) {
-                return cursor.getValue();
-            } else {
-                throw new IllegalArgumentException("Cursor is not an expression. Was " + cursor.getValue().getClass());
-            }
+        private static boolean asExpression(Cursor cursor, Predicate<Expression> expressionPredicate) {
+            return cursor.getValue() instanceof Expression && expressionPredicate.test((Expression) cursor.getValue());
         }
     }
 }
