+++ src/test/java/org/openrewrite/analysis/controlflow/ControlFlowTest.java
@@ -16,8 +16,8 @@
 package org.openrewrite.analysis.controlflow;
 
 import org.junit.jupiter.api.Test;
-import org.openrewrite.Issue;
 import org.openrewrite.DocumentExample;
+import org.openrewrite.Issue;
 import org.openrewrite.test.RecipeSpec;
 import org.openrewrite.test.RewriteTest;
 import org.openrewrite.test.TypeValidation;
