+++ src/test/java/org/openrewrite/analysis/controlflow/GuardTest.java
@@ -16,8 +16,8 @@
 package org.openrewrite.analysis.controlflow;
 
 import org.junit.jupiter.api.Test;
-import org.openrewrite.ExecutionContext;
 import org.openrewrite.DocumentExample;
+import org.openrewrite.ExecutionContext;
 import org.openrewrite.java.JavaIsoVisitor;
 import org.openrewrite.java.tree.Expression;
 import org.openrewrite.java.tree.Statement;
