+++ src/test/java/org/openrewrite/analysis/dataflow/ExternalFlowModelsTest.java
@@ -19,7 +19,6 @@ import org.junit.jupiter.api.Test;
 
 import java.util.HashSet;
 import java.util.Set;
-
 import java.util.stream.Collectors;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
