+++ src/test/java/org/openrewrite/analysis/trait/variable/ParameterTest.java
@@ -19,9 +19,8 @@ import org.junit.jupiter.api.Test;
 import org.openrewrite.Cursor;
 import org.openrewrite.ExecutionContext;
 import org.openrewrite.TreeVisitor;
-import org.openrewrite.analysis.trait.variable.Parameter;
-import org.openrewrite.java.JavaIsoVisitor;
 import org.openrewrite.analysis.trait.util.TraitErrors;
+import org.openrewrite.java.JavaIsoVisitor;
 import org.openrewrite.java.tree.J;
 import org.openrewrite.marker.SearchResult;
 import org.openrewrite.test.RewriteTest;
