+++ src/test/java/org/openrewrite/analysis/trait/variable/VariableTest.java
@@ -17,9 +17,8 @@ package org.openrewrite.analysis.trait.variable;
 
 import org.junit.jupiter.api.Test;
 import org.openrewrite.ExecutionContext;
-import org.openrewrite.analysis.trait.variable.Variable;
-import org.openrewrite.java.JavaIsoVisitor;
 import org.openrewrite.analysis.trait.expr.VarAccess;
+import org.openrewrite.java.JavaIsoVisitor;
 import org.openrewrite.java.tree.J;
 import org.openrewrite.marker.SearchResult;
 import org.openrewrite.test.RecipeSpec;
